corregi las relaciones completamente , elimine la entidad Shoes y deje todo en la entidad Product
