--- karmelshoes/src/main/java/com/karmelshoes/domain/service/IShoesService.java
@@ -1,12 +1,0 @@
-package com.karmelshoes.domain.service;
-
-import com.karmelshoes.persistency.entity.ShoesEntity;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public interface IShoesService {
-
-    List<ShoesEntity> getAll();
-}
