+++ karmelshoes/src/main/java/com/karmelshoes/persistency/entity/ShoppingCartEntity.java
@@ -1,6 +1,5 @@
 package com.karmelshoes.persistency.entity;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 
 import java.util.ArrayList;
@@ -14,21 +13,13 @@ public class ShoppingCartEntity {
     @Column(name = "id", nullable = false)
     private Long id;
 
-    @OneToMany(mappedBy = "shoppingCartEntity", orphanRemoval = false, fetch = FetchType.LAZY)
-    private List<ProductEntity> productEntities = new ArrayList<>();
-
-    @ManyToOne(cascade = CascadeType.PERSIST, fetch = FetchType.LAZY)
+    @ManyToOne(cascade = CascadeType.PERSIST, fetch = FetchType.EAGER)
     @JoinColumn(name = "client_entity_id")
-    @JsonIgnore
     private ClientEntity clientEntity;
 
-    public ClientEntity getClientEntity() {
-        return clientEntity;
-    }
-
-    public void setClientEntity(ClientEntity clientEntity) {
-        this.clientEntity = clientEntity;
-    }
+    @OneToMany(cascade = CascadeType.PERSIST, orphanRemoval = true, fetch = FetchType.EAGER)
+    @JoinColumn(name = "shopping_cart_entity_id")
+    private List<ProductEntity> productEntities = new ArrayList<>();
 
     public List<ProductEntity> getProductEntities() {
         return productEntities;
@@ -36,6 +27,14 @@ public class ShoppingCartEntity {
 
     public void setProductEntities(List<ProductEntity> productEntities) {
         this.productEntities = productEntities;
+    }
+
+    public ClientEntity getClientEntity() {
+        return clientEntity;
+    }
+
+    public void setClientEntity(ClientEntity clientEntity) {
+        this.clientEntity = clientEntity;
     }
 
     public Long getId() {
