--- karmelshoes/src/main/java/com/karmelshoes/persistency/repository/IShoesRepository.java
@@ -1,9 +1,0 @@
-package com.karmelshoes.persistency.repository;
-
-import com.karmelshoes.persistency.entity.ShoesEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface IShoesRepository extends JpaRepository<ShoesEntity, Long> {
-}
