--- karmelshoes/src/main/java/com/karmelshoes/web/controller/ShoesController.java
@@ -1,23 +1,0 @@
-package com.karmelshoes.web.controller;
-
-import com.karmelshoes.domain.service.IShoesService;
-import com.karmelshoes.persistency.entity.ShoesEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/shoes")
-public class ShoesController {
-
-    private final IShoesService iShoesService;
-
-    public ShoesController(IShoesService iShoesService) {
-        this.iShoesService = iShoesService;
-    }
-
-    @GetMapping("/getAll")
-    public List<ShoesEntity> getAll() {
-        return iShoesService.getAll();
-    }
-}
